+++ src/main/lombok/org/heigit/ohsome/ohsomeapi/executor/AggregateRequestExecutor.java
@@ -1,7 +1,6 @@
 package org.heigit.ohsome.ohsomeapi.executor;
 
 import static org.heigit.ohsome.ohsomeapi.utils.GroupByBoundaryGeoJsonGenerator.createGeoJsonFeatures;
-
 import com.opencsv.CSVWriter;
 import java.io.IOException;
 import java.text.DecimalFormat;
@@ -38,7 +37,7 @@ import org.heigit.ohsome.ohsomeapi.output.DefaultAggregationResponse;
 import org.heigit.ohsome.ohsomeapi.output.Description;
 import org.heigit.ohsome.ohsomeapi.output.Metadata;
 import org.heigit.ohsome.ohsomeapi.output.Response;
-import org.heigit.ohsome.ohsomeapi.output.contributions.UsersResult;
+import org.heigit.ohsome.ohsomeapi.output.contributions.ContributionsResult;
 import org.heigit.ohsome.ohsomeapi.output.elements.ElementsResult;
 import org.heigit.ohsome.ohsomeapi.output.groupby.GroupByObject;
 import org.heigit.ohsome.ohsomeapi.output.groupby.GroupByResponse;
@@ -49,8 +48,7 @@ import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.geom.Polygonal;
 
 /**
- * Holds the relevant execution methods for aggregation requests under /elements and
- * /elements/_/groupBy/boundary.
+ * Holds relevant execution methods for various aggregation requests.
  */
 public class AggregateRequestExecutor extends RequestExecutor {
 
@@ -200,12 +198,12 @@ public class AggregateRequestExecutor extends RequestExecutor {
         writer.writeNext(new String[] {elementsResult.getTimestamp(),
             String.valueOf(elementsResult.getValue())});
       }
-    } else if (resultSet instanceof UsersResult[]) {
-      UsersResult[] rs = (UsersResult[]) resultSet;
+    } else if (resultSet instanceof ContributionsResult[]) {
+      ContributionsResult[] rs = (ContributionsResult[]) resultSet;
       writer.writeNext(new String[] {"fromTimestamp", "toTimestamp", "value"}, false);
-      for (UsersResult usersResult : rs) {
-        writer.writeNext(new String[] {usersResult.getFromTimestamp(), usersResult.getToTimestamp(),
-            String.valueOf(usersResult.getValue())});
+      for (ContributionsResult ContributionsResult : rs) {
+        writer.writeNext(new String[] {ContributionsResult.getFromTimestamp(), ContributionsResult.getToTimestamp(),
+            String.valueOf(ContributionsResult.getValue())});
       }
     } else if (resultSet instanceof RatioResult[]) {
       RatioResult[] rs = (RatioResult[]) resultSet;