+++ src/test/java/org/heigit/ohsome/ohsomeapi/controller/PostControllerTest.java
@@ -704,7 +704,6 @@ public class PostControllerTest {
         expectedValue * deltaPercentage);
   }
 
-
   @Test
   public void elementsAreaDensityGroupByTagTest() {
     final double expectedValue = 404281.85;
@@ -815,7 +814,9 @@ public class PostControllerTest {
         expectedValue * deltaPercentage);
   }
 
-  // csv output tests
+  /*
+   * csv output tests
+   */
 
   @Test
   public void elementsLengthCsvTest() throws IOException {
@@ -894,8 +895,6 @@ public class PostControllerTest {
         expectedValue * deltaPercentage);
   }
 
-
-
   @Test
   public void elementsLengthGroupByBoundaryGroupByTagSimpleFeatureCsvTest() throws IOException {
     // expect result to have 1 entry rows with 9 columns
@@ -916,7 +915,6 @@ public class PostControllerTest {
     assertEquals(expectedValue, Double.parseDouble(records.get(0).get("b2_highway=footway")),
         expectedValue * deltaPercentage);
   }
-
 
   @Test
   public void elementsPerimeterCsvTest() throws IOException {
@@ -1079,6 +1077,7 @@ public class PostControllerTest {
   /*
    * filter tests
    */
+  
   @Test
   public void postFilterTest() {
     TestRestTemplate restTemplate = new TestRestTemplate();
@@ -1136,4 +1135,34 @@ public class PostControllerTest {
         restTemplate.postForEntity(server + port + "/elements/count?", map, JsonNode.class);
     assertEquals(null, response.getBody().get("error"));
   }
+
+  /*
+   * /contributions/count tests
+   */
+  
+  @Test
+  public void countContributionsToHeidelbergCastleTest() {
+    TestRestTemplate restTemplate = new TestRestTemplate();
+    MultiValueMap<String, String> map = new LinkedMultiValueMap<>();
+    map.add("bboxes", "8.7137,49.40916,8.71694,49.41198");
+    map.add("time", "2015-01-01,2019-01-01");
+    map.add("filter", "id:way/254154168");
+    ResponseEntity<JsonNode> response =
+        restTemplate.postForEntity(server + port + "/contributions/count", map, JsonNode.class);
+    assertEquals(16, response.getBody().get("result").get(0).get("value").asInt());
+  }
+
+  @Test
+  public void countDensityOfContributionsToShopsInOldtownHeidelbergTest() {
+    TestRestTemplate restTemplate = new TestRestTemplate();
+    MultiValueMap<String, String> map = new LinkedMultiValueMap<>();
+    map.add("bboxes", "8.69282,49.40766,8.71673,49.4133");
+    map.add("time", "2018-01-01,2019-01-01");
+    map.add("filter", "shop=* and type:node");
+    ResponseEntity<JsonNode> response = restTemplate
+        .postForEntity(server + port + "/contributions/count/density", map, JsonNode.class);
+    assertEquals(85.45, response.getBody().get("result").get(0).get("value").asDouble(),
+        deltaPercentage);
+  }
+  
 }